refactor(web): SoC with claude code
App 컴포넌트 내에 있는 비지니스 로직을 다른 파일로 옮겨서 관리하면 좋을 것 같아. 그렇게 해서 App 컴포넌트에서는 필요한 값만 전달 받아서 UI 제어만 하면 좋겠어

useQuery까지 합치는건 별로일까?

hooks 폴더로 옮기는 것보다는 app/index.tsx와 가깝게 두는 게 좋을 것 같은데 어떻게 생각해? app 폴더에 hooks.ts나 helper.ts나 service.ts 등의 이름으로 파일을 만들면 어떨까?
